refactor: :recycle: dateTemplate 변환 함수 dateRange service 로 이동
leaderboard 뿐만 아니라, 여기저기서 활용할 수 있을 것으로 생각이 됐습니다.

- #103
